codeintel: Move gitserver.BranchesContaining to the git package
